api/requirements.txt
django==1.8.12
django-tastypie==0.12.2
django-tastypie-swagger==0.1.4
whitenoise==3.2
raven==5.18.0
python-logstash==0.4.6
gunicorn==19.6.0
django-cors-headers==1.1.0
requests==2.4.2
mysqlclient==1.3.9
